Create Sample Credit Page
